data structure to keep the execution results in
Bug-Url: https://github.com/kerubistan/kerub/issues/146

+++ src/main/kotlin/com/github/K0zka/kerub/data/ispn/EventDaoImpl.kt
@@ -30,7 +30,7 @@ class EventDaoImpl(val cache: AdvancedCache<UUID, Event>) : EventDao {
 		return Search.getQueryFactory(cache)
 				.from(Event::class.java)
 				.orderBy(sort, SortOrder.DESC)
-				.maxResults(limit.toInt())
+				.maxResults(limit)
 				.startOffset(start)
 				.list()
 	}
